+++ migrations/options/src/python/migrate_config.py
@@ -140,10 +140,13 @@ migrations = {
   ('jar-publish', 'repos'): ('publish.jar', 'repos'),
   ('jar-publish', 'publish_extras'): ('publish.jar', 'publish_extras'),
 
-  ('publish', 'restrict_push_branches'): ('publish.jar', 'restrict_push_branches'),
+  ('publish', 'individual_plugins'): ('publish.jar', 'individual_plugins'),
+  ('publish', 'ivy_settings'): ('publish.jar', 'ivy_settings'),
   ('publish', 'jvm_options'): ('publish.jar', 'jvm_options'),
-  ('publish', 'repos'): ('publish.jar', 'repos'),
   ('publish', 'publish_extras'): ('publish.jar', 'publish_extras'),
+  ('publish', 'push_postscript'): ('publish.jar', 'push_postscript'),
+  ('publish', 'repos'): ('publish.jar', 'repos'),
+  ('publish', 'restrict_push_branches'): ('publish.jar', 'restrict_push_branches'),
 
   # Three changes are pertinent to migrate 'ide' to both idea and & eclipse. I tried to capture
   # that in notes
@@ -225,6 +228,9 @@ migrations = {
   ('ragel-gen', 'version'): ('gen.ragel', 'version'),
 
   ('prepare-resources', 'confs'): ('resources.prepare', 'confs'),
+
+  ('compile.scala', 'runtime-deps'): ('scala-platform', 'runtime'),
+  ('compile.scala', 'scalac'): ('scala-platform', 'scalac'),
 }
 
 ng_daemons_note = ('The global "ng_daemons" option has been replaced by a "use_nailgun" option '