+++ src/scala/org/pantsbuild/zinc/Inputs.scala
@@ -5,146 +5,169 @@
 package org.pantsbuild.zinc
 
 import java.io.File
-import java.util.{List => JList, Map => JMap}
-
+import java.util.{ List => JList, Map => JMap }
+import sbt.compiler.IC
+import sbt.inc.{ Analysis, Locate }
 import sbt.Path._
-import sbt.inc.{Analysis, Locate}
-import xsbti.compile.CompileOrder
-
 import scala.collection.JavaConverters._
+import xsbti.compile.CompileOrder
 
 /**
  * All inputs for a compile run.
  */
 case class Inputs(
-    classpath: Seq[File],
-    sources: Seq[File],
-    classesDirectory: File,
-    scalacOptions: Seq[String],
-    javacOptions: Seq[String],
-    cacheFile: File,
-    analysisMap: Map[File, Analysis],
-    forceClean: Boolean,
-    definesClass: File => String => Boolean,
-    javaOnly: Boolean,
-    compileOrder: CompileOrder,
-    incOptions: IncOptions,
-    outputRelations: Option[File],
-    outputProducts: Option[File],
-    mirrorAnalysis: Boolean)
+  classpath: Seq[File],
+  sources: Seq[File],
+  classesDirectory: File,
+  scalacOptions: Seq[String],
+  javacOptions: Seq[String],
+  cacheFile: File,
+  analysisMap: Map[File, Analysis],
+  forceClean: Boolean,
+  definesClass: File => String => Boolean,
+  javaOnly: Boolean,
+  compileOrder: CompileOrder,
+  incOptions: IncOptions,
+  outputRelations: Option[File],
+  outputProducts: Option[File],
+  mirrorAnalysis: Boolean)
 
 object Inputs {
   /**
    * Create inputs based on command-line settings.
    */
-  def apply(settings: Settings): Inputs = {
+  def apply(log: sbt.Logger, settings: Settings): Inputs = {
+    import settings._
     inputs(
-      settings.classpath,
-      settings.sources,
-      settings.classesDirectory,
-      settings.scalacOptions,
-      settings.javacOptions,
-      settings.analysis.cache,
-      settings.analysis.cacheMap,
-      settings.analysis.forceClean,
-      settings.javaOnly,
-      settings.compileOrder,
-      settings.incOptions,
-      settings.analysis.outputRelations,
-      settings.analysis.outputProducts,
-      settings.analysis.mirrorAnalysis)
-  }
-  @deprecated("Use the variant that takes `incOptions` parameter, instead.", "0.3.5.3")
-  def create(
-      classpath: JList[File],
-      sources: JList[File],
-      classesDirectory: File,
-      scalacOptions: JList[String],
-      javacOptions: JList[String],
-      analysisCache: File,
-      analysisMap: JMap[File, File],
-      compileOrder: String,
-      mirrorAnalysisCache: Boolean): Inputs =
-    create(classpath, sources, classesDirectory, scalacOptions, javacOptions,
-      analysisCache, analysisMap, compileOrder, IncOptions(), mirrorAnalysisCache)
-  /**
-   * Java API for creating Inputs.
-   */
-  def create(
-      classpath: JList[File],
-      sources: JList[File],
-      classesDirectory: File,
-      scalacOptions: JList[String],
-      javacOptions: JList[String],
-      analysisCache: File,
-      analysisMap: JMap[File, File],
-      compileOrder: String,
-      incOptions: IncOptions,
-      mirrorAnalysisCache: Boolean): Inputs =
-    inputs(
-      classpath.asScala,
-      sources.asScala,
+      log,
+      classpath,
+      sources,
       classesDirectory,
-      scalacOptions.asScala,
-      javacOptions.asScala,
-      Option(analysisCache),
-      analysisMap.asScala.toMap,
-      forceClean = false,
-      javaOnly = false,
-      Settings.compileOrder(compileOrder),
+      scalacOptions,
+      javacOptions,
+      analysis.cache,
+      analysis.cacheMap,
+      analysis.forceClean,
+      javaOnly,
+      compileOrder,
       incOptions,
-      outputRelations = None,
-      outputProducts = None,
-      mirrorAnalysis = mirrorAnalysisCache
-    )
+      analysis.outputRelations,
+      analysis.outputProducts,
+      analysis.mirrorAnalysis)
+  }
+
+  /** An overridden definesClass to use analysis for an input directory if it is available. */
+  def definesClass(log: sbt.Logger, analysisMap: Map[File, Analysis], entry: File): String => Boolean =
+    analysisMap.get(entry).map { analysis =>
+      log.debug(s"Hit analysis cache for class definitions with ${entry}")
+      (s: String) => analysis.relations.definesClass(s).nonEmpty
+    }.getOrElse {
+      Locate.definesClass(entry)
+    }
+
   /**
    * Create normalised and defaulted Inputs.
    */
   def inputs(
-      classpath: Seq[File],
-      sources: Seq[File],
-      classesDirectory: File,
-      scalacOptions: Seq[String],
-      javacOptions: Seq[String],
-      analysisCache: Option[File],
-      analysisCacheMap: Map[File, File],
-      forceClean: Boolean,
-      javaOnly: Boolean,
-      compileOrder: CompileOrder,
-      incOptions: IncOptions,
-      outputRelations: Option[File],
-      outputProducts: Option[File],
-      mirrorAnalysis: Boolean): Inputs = {
-    val normalise: File => File = {
-      _.getAbsoluteFile
-    }
-    val cp = classpath map normalise
-    val srcs = sources map normalise
-    val classes = normalise(classesDirectory)
-    val cacheFile = normalise(analysisCache.getOrElse(defaultCacheLocation(classesDirectory)))
+    log: sbt.Logger,
+    classpath: Seq[File],
+    sources: Seq[File],
+    classesDirectory: File,
+    scalacOptions: Seq[String],
+    javacOptions: Seq[String],
+    analysisCache: Option[File],
+    analysisCacheMap: Map[File, File],
+    forceClean: Boolean,
+    javaOnly: Boolean,
+    compileOrder: CompileOrder,
+    incOptions: IncOptions,
+    outputRelations: Option[File],
+    outputProducts: Option[File],
+    mirrorAnalysis: Boolean): Inputs =
+  {
+    val normalise: File => File = { _.getAbsoluteFile }
+    val cp               = classpath map normalise
+    val srcs             = sources map normalise
+    val classes          = normalise(classesDirectory)
+    val cacheFile        = normalise(analysisCache.getOrElse(defaultCacheLocation(classesDirectory)))
     val upstreamAnalysis = analysisCacheMap map { case (k, v) => (normalise(k), normalise(v)) }
-    val analysisMap = (cp map { file => (file, analysisFor(file, classes, upstreamAnalysis)) }).toMap
-    val incOpts = updateIncOptions(incOptions, classesDirectory, normalise)
-    val printRelations = outputRelations map normalise
-    val printProducts = outputProducts map normalise
+    // Use only existing upstream analysis files for class lookups.
+    val validUpstreamAnalysis =
+      upstreamAnalysis.flatMap {
+        case (k, _) if k == classes =>
+          // ignore our own analysis
+          None
+        case (k, v) =>
+          // use analysis only if it was valid/non-empty
+          Compiler.analysisOption(v).map { analysis =>
+            k -> analysis
+          }
+      }
+    val analysisMap      = (cp map { file => (file, allAnalysisFor(file, classes, upstreamAnalysis)) }).toMap
+    val incOpts          = updateIncOptions(incOptions, classesDirectory, normalise)
+    val printRelations   = outputRelations map normalise
+    val printProducts    = outputProducts map normalise
     new Inputs(
-      cp, srcs, classes, scalacOptions, javacOptions, cacheFile, analysisMap, forceClean, Locate.definesClass,
+      cp, srcs, classes, scalacOptions, javacOptions, cacheFile, analysisMap, forceClean, definesClass(log, validUpstreamAnalysis, _),
       javaOnly, compileOrder, incOpts, printRelations, printProducts, mirrorAnalysis
     )
   }
+
+  /**
+   * Java API for creating Inputs.
+   */
+  def create(
+    log: sbt.Logger,
+    classpath: JList[File],
+    sources: JList[File],
+    classesDirectory: File,
+    scalacOptions: JList[String],
+    javacOptions: JList[String],
+    analysisCache: File,
+    analysisMap: JMap[File, File],
+    compileOrder: String,
+    incOptions: IncOptions,
+    mirrorAnalysisCache: Boolean): Inputs =
+  inputs(
+    log,
+    classpath.asScala,
+    sources.asScala,
+    classesDirectory,
+    scalacOptions.asScala,
+    javacOptions.asScala,
+    Option(analysisCache),
+    analysisMap.asScala.toMap,
+    forceClean = false,
+    javaOnly = false,
+    Settings.compileOrder(compileOrder),
+    incOptions,
+    outputRelations = None,
+    outputProducts = None,
+    mirrorAnalysis = mirrorAnalysisCache
+  )
+
+  @deprecated("Use the variant that takes `incOptions` parameter, instead.", "0.3.5.3")
+  def create(
+    log: sbt.Logger,
+    classpath: JList[File],
+    sources: JList[File],
+    classesDirectory: File,
+    scalacOptions: JList[String],
+    javacOptions: JList[String],
+    analysisCache: File,
+    analysisMap: JMap[File, File],
+    compileOrder: String,
+    mirrorAnalysisCache: Boolean): Inputs =
+    create(log, classpath, sources, classesDirectory, scalacOptions, javacOptions,
+      analysisCache, analysisMap, compileOrder, IncOptions(), mirrorAnalysisCache)
+
   /**
    * By default the cache location is relative to the classes directory (for example, target/classes/../cache/classes).
    */
   def defaultCacheLocation(classesDir: File) = {
     classesDir.getParentFile / "cache" / classesDir.getName
   }
-  /**
-   * Get the analysis for a compile run, based on a classpath entry.
-   * If not cached in memory, reads from the cache file.
-   */
-  def analysisFor(file: File, exclude: File, mapped: Map[File, File]): Analysis = {
-    cacheFor(file, exclude, mapped) map Compiler.analysis getOrElse Analysis.Empty
-  }
+
   /**
    * Get the possible cache location for a classpath entry. Checks the upstream analysis map
    * for the cache location, otherwise uses the default location for output directories.
@@ -154,6 +177,15 @@ object Inputs {
       if (file.isDirectory && file != exclude) Some(defaultCacheLocation(file)) else None
     }
   }
+
+  /**
+   * Get the analysis for a compile run, based on a classpath entry.
+   * If not cached in memory, reads from the cache file, or creates empty analysis.
+   */
+  def allAnalysisFor(file: File, exclude: File, mapped: Map[File, File]): Analysis = {
+    cacheFor(file, exclude, mapped) map Compiler.analysis getOrElse Analysis.Empty
+  }
+
   /**
    * Normalise files and default the backup directory.
    */
@@ -210,34 +242,35 @@ object Inputs {
    * Debug output for inputs.
    */
   def show(inputs: Inputs, output: String => Unit): Unit = {
+    import inputs._
 
     val incOpts = Seq(
-      "transitive step" -> inputs.incOptions.transitiveStep,
-      "recompile all fraction" -> inputs.incOptions.recompileAllFraction,
-      "debug relations" -> inputs.incOptions.relationsDebug,
-      "debug api" -> inputs.incOptions.apiDebug,
-      "api dump" -> inputs.incOptions.apiDumpDirectory,
-      "api diff context size" -> inputs.incOptions.apiDiffContextSize,
-      "transactional" -> inputs.incOptions.transactional,
-      "backup directory" -> inputs.incOptions.backup,
-      "recompile on macro def" -> inputs.incOptions.recompileOnMacroDef,
-      "name hashing" -> inputs.incOptions.nameHashing
+      "transitive step"        -> incOptions.transitiveStep,
+      "recompile all fraction" -> incOptions.recompileAllFraction,
+      "debug relations"        -> incOptions.relationsDebug,
+      "debug api"              -> incOptions.apiDebug,
+      "api dump"               -> incOptions.apiDumpDirectory,
+      "api diff context size"  -> incOptions.apiDiffContextSize,
+      "transactional"          -> incOptions.transactional,
+      "backup directory"       -> incOptions.backup,
+      "recompile on macro def" -> incOptions.recompileOnMacroDef,
+      "name hashing"           -> incOptions.nameHashing
     )
 
     val values = Seq(
-      "classpath" -> inputs.classpath,
-      "sources" -> inputs.sources,
-      "output directory" -> inputs.classesDirectory,
-      "scalac options" -> inputs.scalacOptions,
-      "javac options" -> inputs.javacOptions,
-      "cache file" -> inputs.cacheFile,
-      "analysis map" -> inputs.analysisMap,
-      "force clean" -> inputs.forceClean,
-      "java only" -> inputs.javaOnly,
-      "compile order" -> inputs.compileOrder,
+      "classpath"                    -> classpath,
+      "sources"                      -> sources,
+      "output directory"             -> classesDirectory,
+      "scalac options"               -> scalacOptions,
+      "javac options"                -> javacOptions,
+      "cache file"                   -> cacheFile,
+      "analysis map"                 -> analysisMap,
+      "force clean"                  -> forceClean,
+      "java only"                    -> javaOnly,
+      "compile order"                -> compileOrder,
       "incremental compiler options" -> incOpts,
-      "output relations" -> inputs.outputRelations,
-      "output products" -> inputs.outputProducts)
+      "output relations"             -> outputRelations,
+      "output products"              -> outputProducts)
 
     Util.show(("Inputs", values), output)
   }