+++ tests/python/pants_test/backend/jvm/tasks/test_jvm_binary_integration.py
@@ -16,12 +16,12 @@ class JvmBinaryIntegrationTest(PantsRunIntegrationTest):
 
   def test_manifest_entries(self):
     self.assert_success(self.run_pants(['clean-all']))
-    args = ['binary', 'testprojects/src/java/org/pantsbuild/testproject/manifest']
+    args = ['binary', 'testprojects/src/java/org/pantsbuild/testproject/manifest:manifest-with-source']
     pants_run = self.run_pants(args, {})
     self.assert_success(pants_run)
 
     out_path = os.path.join(get_buildroot(), 'dist')
-    java_run = subprocess.Popen(['java', '-cp', 'manifest.jar',
+    java_run = subprocess.Popen(['java', '-cp', 'manifest-with-source.jar',
                                  'org.pantsbuild.testproject.manifest.Manifest'],
                                 stdout=subprocess.PIPE,
                                 cwd=out_path)
@@ -29,3 +29,35 @@ class JvmBinaryIntegrationTest(PantsRunIntegrationTest):
     java_out = java_run.stdout.read()
     self.assertEquals(java_retcode, 0)
     self.assertIn('Hello World!  Version: 1.2.3', java_out)
+
+  def test_manifest_entries_no_source(self):
+    self.assert_success(self.run_pants(['clean-all']))
+    args = ['binary', 'testprojects/src/java/org/pantsbuild/testproject/manifest:manifest-no-source']
+    pants_run = self.run_pants(args, {})
+    self.assert_success(pants_run)
+
+    out_path = os.path.join(get_buildroot(), 'dist')
+    java_run = subprocess.Popen(['java', '-cp', 'manifest-no-source.jar',
+                                 'org.pantsbuild.testproject.manifest.Manifest'],
+                                stdout=subprocess.PIPE,
+                                cwd=out_path)
+    java_retcode = java_run.wait()
+    java_out = java_run.stdout.read()
+    self.assertEquals(java_retcode, 0)
+    self.assertIn('Hello World!  Version: 4.5.6', java_out)
+
+  def test_manifest_entries_bundle(self):
+    self.assert_success(self.run_pants(['clean-all']))
+    args = ['bundle', 'testprojects/src/java/org/pantsbuild/testproject/manifest:manifest-app']
+    pants_run = self.run_pants(args, {})
+    self.assert_success(pants_run)
+
+    out_path = os.path.join(get_buildroot(), 'dist', 'manifest-app-bundle')
+    java_run = subprocess.Popen(['java', '-cp', 'manifest-no-source.jar',
+                                 'org.pantsbuild.testproject.manifest.Manifest'],
+                                stdout=subprocess.PIPE,
+                                cwd=out_path)
+    java_retcode = java_run.wait()
+    java_out = java_run.stdout.read()
+    self.assertEquals(java_retcode, 0)
+    self.assertIn('Hello World!  Version: 4.5.6', java_out)